Create a development app runner script
This script allows the app to be ran in a terminal window with
detailed logging and automatic reloading. Detailed logging makes
it easier to see what is actually happening between the browser and
the shiny app. Automatic reloading refreshes the browser every time
a file is saved in the application which can improve the efficiency
of iterative development.