Add cell type and gene expression plots
The MARVEL plot functions do not provide the ability to segregate
umap data by wildtype and mutant. As a result, code had to be
written to separate the gene expression data that is used by
PlotValues.PCA.Gene.10x. The code located in the MARVEL github
repo was used to determine which data needed filtering.

Also, MARVEL does not provide a way to specify a title for plots.
The code located in the MARVEL github repo was used to determine
where plots were stored on the MARVEL object so a title could be
added.