Improve UI interaction and performance
Added help text, improved prompts, and hide/show plots and
selectors based on user selections to improve the UI interaction.

The plots generate very slowly. When the UI automatically updates
as selections of genes and splice junctions occur the slowness is
frustrating.  A button was added to show plots once the user has
completed their selection.  The button did not completely solve the
performance problems because selecting different splice junctions
for the same gene would cause the genes to replot unnecessarily.
Caching was added to the plots to prevent unneeded recalculations.