Remove log2 transform on gene expression plots
The gene expression data is already normalized so it does not need
a log2 tranformation.

The MARVEL gene expression plotting function does not work correctly
when log2.transform = FALSE so the code had to be monkey patched
to fix the defect.  The file containing the function is in the
MARVEL folder and has been modified to correct the bug.

The following pull request was submitted to the MARVEL project on
Github.  Once the pull request has been merged and released the
monkey patch can be removed.

https://github.com/wenweixiong/MARVEL/pull/41