Pre-render the cell type plot
The cell type plot is slow and requires all the MARVEL data
to be loaded into memory.  The plot never changes since it
is showing all the cells so pre-rendering is possible and
solves the memory and performance problems.

The R/plot.R script contains the function to pre-render plots.

Example in R Console:

> source("R/plot.R")
> prerenderPlots("data/setbp1_marvel_alligned_sans_gtf.rds")