Add rsconnect .rscignore file
The rsconnect library is used to upload Shiny apps to shinyapps.io
and the .rscignore file works like the .gitignore file.  This
prevents large source data files from being uploaded and exceeding
the shinyapps.io bundle size limit.  The ignoring is less robust
than .gitignore in that it will only ignore files and directories
in the application root directory.