Expand maximum bundle size for deployment
The default maximum bundle size for rsconnect deployment is 3 Gb. This
limit was exceeded and deployments failed. Our host, shinyapps.io,
has a 5 Gb maximum bundle size which is sufficient for our deployments.
The maximum can be modified in the .rsconnect_profile so a profile was
created with a maximum of 5 Gb.