Prevent plot rendering when the tissue changes
When the tissue changed the current plots would rerender making
the tissue change very slow. Plots did not need to rerender until
the gene and/or splice junction is selected. The rerender seemed
fine when I thought the meta data for brain and kidney were the same.
Now that they are determined to be different, those options need to
be reselected before plots can be rerendered on a tissue change.